% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-academic_year.R
\name{academic_year}
\alias{academic_year}
\title{Internal vctrs methods}
\usage{
academic_year(
  x = integer(),
  boundary = getOption("academic_year_boundary", as.Date("2020-08-01"))
)
}
\arguments{
\item{x}{\itemize{
\item For \code{academic_year()}: A numeric vector
\itemize{
\item For \code{as_academic_year()}: An object to coerce
\item For \code{is_academic_year()}: An object to test
}
}}

\item{boundary}{A boundary date where the new academic year should begin}
}
\description{
Internal vctrs methods

Academic year vector
}
\keyword{internal}
