% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-academic_year.R
\name{as.POSIXct.academic_year}
\alias{as.POSIXct.academic_year}
\title{#' @export
as.Date.academic_year <- function(x, ...) {
if (length(x) == 0L) {
return(as.Date(NULL))
}
boundary <- strftime(attr(x, "boundary"), "-\%m-\%d")
as.Date(paste0(as.integer(x), boundary))
}}
\usage{
\method{as.POSIXct}{academic_year}(x, ...)
}
\description{
#' @export
as.Date.academic_year <- function(x, ...) {
if (length(x) == 0L) {
return(as.Date(NULL))
}
boundary <- strftime(attr(x, "boundary"), "-\%m-\%d")
as.Date(paste0(as.integer(x), boundary))
}
}
