% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-financial_year.R
\name{financial_year}
\alias{financial_year}
\title{Internal vctrs methods}
\usage{
financial_year(
  x = integer(),
  boundary = getOption("financial_year_boundary", as.Date("2020-04-01"))
)
}
\arguments{
\item{x}{\itemize{
\item For \code{financial_year()}: A numeric vector
\itemize{
\item For \code{as_financial_year()}: An object to coerce
\item For \code{is_financial_year()}: An object to test
}
}}

\item{boundary}{A boundary date where the new financial year should begin}
}
\description{
Internal vctrs methods

financial year vector
}
\keyword{internal}
