% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-academic_year.R
\name{set_year_boundary}
\alias{set_year_boundary}
\alias{year_boundary<-}
\title{Set the boundary when a new year begins}
\usage{
set_year_boundary(x, y)

year_boundary(x, y) <- value
}
\arguments{
\item{x}{An academic year object}

\item{y}{A date}
}
\description{
Set the boundary when a new year begins
}
